Replace PlayBuf with XPlayBuf
This UGen helps avoid pops and clicks at loop points by
cross-fading. It also performs the samplerate to rate scaling inside
the UGen so there's no need for the call to `BufRateScale`.

0.5 second xFadeTime seems to help with clicks and pops on the sound
files I've tried.

For now, while developing on the mac, I've installed XPlayBuf manually
using the instructions in the README. If I want to share this script
on the norns, I'll need to do something more sophisticated. Zack's
tapedeck script[1] handles this neatly.

[1] https://github.com/schollz/tapedeck/blob/main/tapedeck.lua#L40